feat: 댓글 삭제 시 댓글 작성자 or 관리자 검증
관리자는 서비스 관리 목적으로 모든 댓글 삭제 가능

+++ src/main/java/com/ll/JollyJourney/domain/journalcomment/service/JournalCommentService.java
@@ -72,8 +72,18 @@ public class JournalCommentService {
     }
 
     @Transactional
-    public void deleteComment(Long commentId) {
+    public void deleteComment(Long commentId, Authentication authentication) {
+        JournalComment journalComment = journalCommentRepository.findById(commentId)
+                .orElseThrow(() -> new IllegalArgumentException("해당 댓글 없음"));
 
+        CustomUserDetails userDetails = (CustomUserDetails) authentication.getPrincipal();
+        Long currentUserId = userDetails.getMemberId();
+        boolean isAdmin = authentication.getAuthorities().stream()
+                .anyMatch(grantedAuthority -> grantedAuthority.getAuthority().equals("ROLE_ADMIN"));
+        
+        if (!journalComment.getMember().getUserId().equals(currentUserId) && !isAdmin) {
+            throw new AccessDeniedException("댓글 작성자 또는 관리자만 삭제할 수 있습니다.");
+        }
         journalCommentRepository.deleteById(commentId);
     }
 }
